+++ CHANGELOG.md
@@ -8,6 +8,8 @@ semantic-release.
 
 ## [Unreleased]
 
+## [1.1.0] - 2026-07-24
+
 ## [1.0.1] - 2026-07-24
 
 ### Added
@@ -300,7 +302,8 @@ Release version.
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.1.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/ryancswallace/jobman/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/ryancswallace/jobman/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/ryancswallace/jobman/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/ryancswallace/jobman/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/ryancswallace/jobman/compare/v0.9.0...v1.0.0
 [0.9.0]: https://github.com/ryancswallace/jobman/compare/v0.8.4...v0.9.0